refactor(system-edit): move Parse pill to the page header, drop System details title, high-contrast form labels
