+++ js/app/main.js
@@ -1,7 +1,7 @@
 // Entry point for the app page: bootstraps routing and search wiring.
 // (Replaces the inline bootstrap that lived at the top/bottom of app.js.)
-import { route } from './router.js?v=51784fcf';
-import { wireSearch } from './components/search.js?v=ec9def6e';
+import { route } from './router.js?v=51473916';
+import { wireSearch } from './components/search.js?v=fb715a40';
 
 window.addEventListener('hashchange', () => route());
 window.addEventListener('popstate', () => route());
