+++ profile.html
@@ -12,7 +12,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=86052a10">
 <link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
-<link rel="stylesheet" href="css/profile/profile.css?v=1ddc413a">
+<link rel="stylesheet" href="css/profile/profile.css?v=fea192b9">
 </head>
 <body>
