+++ system-edit.html
@@ -12,7 +12,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=86052a10">
 <link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
-<link rel="stylesheet" href="css/profile/profile.css?v=1ddc413a">
+<link rel="stylesheet" href="css/profile/profile.css?v=fea192b9">
 </head>
 <body>
 
@@ -31,9 +31,12 @@
       </a>
     </div>
 
-    <div class="page-header" style="margin-bottom:20px">
-      <div class="eyebrow" id="page-eyebrow">Edit System</div>
-      <div class="page-title" id="page-title">Loading...</div>
+    <div class="page-header sys-page-header" style="margin-bottom:20px">
+      <div class="sys-page-header-titles">
+        <div class="eyebrow" id="page-eyebrow">Edit System</div>
+        <div class="page-title" id="page-title">Loading...</div>
+      </div>
+      <button type="button" class="profile-parse-btn" id="steam-parse-open" hidden>Parse from Steam info</button>
     </div>
 
     <div id="auth-gate" hidden>
@@ -47,11 +50,7 @@
     </div>
 
     <form id="system-form" class="profile-section" hidden>
-      <div class="sys-form-head">
-        <div class="profile-section-title" id="form-section-title" style="margin:0">System details</div>
-        <button type="button" class="profile-parse-btn" id="steam-parse-open">Parse from Steam info</button>
-      </div>
-      <p class="profile-field-hint" style="margin:0 0 16px">Fill the fields below, or paste your Steam System Information to autofill them.</p>
+      <p class="profile-field-hint" style="margin:0 0 16px">Fill the fields below, or use Parse from Steam info above to autofill them.</p>
       <div class="profile-field">
         <label class="profile-field-label" for="sys-label">Label</label>
         <input type="text" id="sys-label" class="profile-select" style="width:100%" maxlength="80" placeholder="e.g. Desktop RTX 4070">
@@ -130,6 +129,6 @@
   </div>
 </div>
 
-<script type="module" src="js/profile/system-edit.js?v=9dc0b1b6"></script>
+<script type="module" src="js/profile/system-edit.js?v=6ce4b31b"></script>
 </body>
 </html>
